+++ examples/demo-app/src/cloud-providers/foursquare/foursquare-provider.js
@@ -73,6 +73,7 @@ export default class FoursquareProvider extends Provider {
       clientId,
       scope: FOURSQUARE_AUTH_SCOPE,
       authorizationParams: {
+        prompt: 'login',
         redirect_uri: window.location.origin,
         audience: FOURSQUARE_AUTH_AUDIENCE
       },
@@ -158,7 +159,7 @@ export default class FoursquareProvider extends Provider {
       }
     );
     const data = await response.json();
-    return data.items.map(map => convertFSQModelToMapItem(map, `${this.apiURL}/v1/maps`));
+    return (data?.items || []).map(map => convertFSQModelToMapItem(map, `${this.apiURL}/v1/maps`));
   }
 
   async downloadMap(loadParams) {