+++ src/layers/src/layer-utils.ts
@@ -5,12 +5,16 @@ import * as arrow from 'apache-arrow';
 import {Feature, BBox} from 'geojson';
 import {getGeoMetadata} from '@loaders.gl/gis';
 
+import {GEOARROW_EXTENSIONS, GEOARROW_METADATA_KEY} from '@kepler.gl/constants';
+import {KeplerTable} from '@kepler.gl/table';
 import {
   Field,
   ProtoDatasetField,
   FieldPair,
   SupportedColumnMode,
-  LayerColumn
+  LayerColumn,
+  LayerColumns,
+  RGBColor
 } from '@kepler.gl/types';
 import {DataContainerInterface, ArrowDataContainer} from '@kepler.gl/utils';
 import {
@@ -19,7 +23,6 @@ import {
   BinaryGeometriesFromArrowOptions,
   updateBoundsFromGeoArrowSamples
 } from '@loaders.gl/arrow';
-import {EXTENSION_NAME} from '@kepler.gl/deckgl-arrow-layers';
 
 import {WKBLoader} from '@loaders.gl/wkt';
 import {geojsonToBinary} from '@loaders.gl/gis';
@@ -33,7 +36,23 @@ import {
 
 import {DeckGlGeoTypes, GeojsonDataMaps} from './geojson-layer/geojson-utils';
 
-export function assignPointPairToLayerColumn(pair: FieldPair, hasAlt: boolean) {
+export type FindDefaultLayerProps = {
+  label: string;
+  color?: RGBColor;
+  isVisible?: boolean;
+  columns?: Record<string, LayerColumn>;
+};
+
+export type FindDefaultLayerPropsReturnValue = {
+  /** Layer props to create layers by default when a dataset is added */
+  props: FindDefaultLayerProps[];
+  /** layer props of possible alternative layer configurations, not created by default */
+  altProps?: FindDefaultLayerProps[];
+  /** Already found layer configurations */
+  foundLayers?: (FindDefaultLayerProps & {type: string})[];
+};
+
+export function assignPointPairToLayerColumn(pair: FieldPair, hasAlt: boolean): Record<string, LayerColumn> {
   const {lat, lng, altitude} = pair.pair;
   if (!hasAlt) {
     return {lat, lng};
@@ -159,7 +178,7 @@ export function getGeojsonLayerMetaFromArrow({
   geoField: ProtoDatasetField;
   chunkIndex?: number;
 }): GeojsonLayerMetaProps {
-  const encoding = geoField?.metadata?.get('ARROW:extension:name');
+  const encoding = geoField?.metadata?.get(GEOARROW_METADATA_KEY);
   const options: BinaryGeometriesFromArrowOptions = {
     ...(chunkIndex !== undefined && chunkIndex >= 0
       ? {
@@ -172,7 +191,7 @@ export function getGeojsonLayerMetaFromArrow({
   };
 
   // getBinaryGeometriesFromArrow doesn't support geoarrow.wkb
-  if (encoding === EXTENSION_NAME.WKB) {
+  if (encoding === GEOARROW_EXTENSIONS.WKB) {
     return getBinaryGeometriesFromWKBArrow(geoColumn, options);
   }
 
@@ -220,7 +239,7 @@ export function getHoveredObjectFromArrow(
     const rawGeometry = col?.get(objectInfo.index);
 
     const field = fieldAccessor(dataContainer);
-    const encoding = field?.metadata?.get('ARROW:extension:name');
+    const encoding = field?.metadata?.get(GEOARROW_METADATA_KEY);
 
     const hoveredFeature = parseGeometryFromArrow(rawGeometry, encoding);
 
@@ -265,7 +284,7 @@ export function getGeoArrowPointFields(fields: Field[]): Field[] {
   return fields.filter(field => {
     return (
       field.type === 'geoarrow' &&
-      field.metadata?.get('ARROW:extension:name') === EXTENSION_NAME.POINT
+      field.metadata?.get(GEOARROW_METADATA_KEY) === GEOARROW_EXTENSIONS.POINT
     );
   });
 }
@@ -291,7 +310,7 @@ export function createGeoArrowPointVector(
   const precision = 2;
 
   const metadata = new Map();
-  metadata.set('ARROW:extension:name', EXTENSION_NAME.POINT);
+  metadata.set(GEOARROW_METADATA_KEY, GEOARROW_EXTENSIONS.POINT);
 
   const childField = new arrow.Field('xyz', new arrow.Float(precision), false, metadata);
   const fixedSizeList = new arrow.FixedSizeList(numCoords, childField);
@@ -410,4 +429,62 @@ export function getBoundsFromArrowMetadata(
   }
 
   return false;
+}
+
+/**
+ * Finds and returns the first satisfied column mode based on the provided columns and fields.
+ * @param supportedColumnModes - An array of supported column modes to check.
+ * @param columns - The available columns.
+ * @param fields - Optional table fields to be used for extra verification.
+ * @returns The first column mode that satisfies the required conditions, or undefined if none match.
+ */
+export function getSatisfiedColumnMode(
+  columnModes: SupportedColumnMode[] | null,
+  columns: LayerColumns | undefined,
+  fields?: KeplerTable['fields']
+): SupportedColumnMode | undefined {
+  return columnModes?.find(mode => {
+    return mode.requiredColumns?.every(requriedCol => {
+      const column = columns?.[requriedCol];
+      if (column?.value) {
+        if (mode.verifyField && fields?.[column.fieldIdx]) {
+          const field = fields[column.fieldIdx];
+          return mode.verifyField(field);
+        }
+        return true;
+      }
+      return false;
+    });
+  });
+}
+
+/**
+ * Returns true if the field is of geoarrow point format.
+ * @param field A field.
+ * @returns Returns true if the field is of geoarrow point format.
+ */
+export function isGeoArrowPointField(field: Field) {
+  return (
+    field.type === 'geoarrow' &&
+    field.metadata?.get(GEOARROW_METADATA_KEY) === GEOARROW_EXTENSIONS.POINT
+  );
+}
+
+/**
+ * Create default geoarrow column props based on the dataset.
+ * @param dataset A dataset to create layer props from.
+ * @returns  geoarrow column props.
+ */
+export function getGeoArrowPointLayerProps(dataset: KeplerTable) {
+  const {label} = dataset;
+  const altProps: FindDefaultLayerProps[] = [];
+  dataset.fields.forEach(field => {
+    if (isGeoArrowPointField(field)) {
+      altProps.push({
+        label: (typeof label === 'string' && label.replace(/\.[^/.]+$/, '')) || field.name,
+        columns: {geoarrow: {value: field.name, fieldIdx: field.fieldIdx}}
+      });
+    }
+  });
+  return altProps;
 }